started restructure
