+++ .config/nvim/init.lua
@@ -1,4 +1,5 @@
-require('custom')
+require('remaps')
+require('settings')
 
 local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
 if not vim.loop.fs_stat(lazypath) then
@@ -14,5 +15,5 @@ end
 vim.opt.rtp:prepend(lazypath)
 
 require("lazy").setup({
-  {import = 'custom.plugins'}
+  {spec = 'plugins'}
 })
